Extract FilteringIterator from AgeOffIterator
